requirements.txt
streamlit
groq
langchain
python-dotenv